Added colors, font and layout structures
