+++ src/app/(pages)/(dashboard)/layout.tsx
@@ -1,0 +1,7 @@
+export default function DashboardLayout({ children }: { children: React.ReactNode }) {
+    return (
+        <main>
+            {children}
+        </main>
+    );
+};
