+++ src/app/(pages)/(dashboard)/page.tsx
@@ -1,0 +1,5 @@
+export default function Dashboard() {
+  return (
+    <h2>Dashboard</h2>
+  );
+}
