+++ src/app/(pages)/auth/layout.tsx
@@ -1,0 +1,7 @@
+export default function AuthLayout({ children }: { children: React.ReactNode }) {
+    return (
+        <main>
+            {children}
+        </main>
+    );
+};
