+++ src/app/globals.css
@@ -1,3 +1,4 @@
+@import url('https://fonts.cdnfonts.com/css/tt-firs-neue-trl');
 @tailwind base;
 @tailwind components;
 @tailwind utilities;
@@ -17,5 +18,5 @@
 body {
   color: var(--foreground);
   background: var(--background);
-  font-family: Arial, Helvetica, sans-serif;
+  font-family: 'TT Firs Neue Trl', sans-serif;
 }
